Revert "move runResource to use-resource file"
This reverts commit 8aad31a473e03334b5fcdd8c4028526425fe7953.
