+++ packages/qwik/src/core/v2/shared/scheduler.ts
@@ -84,13 +84,14 @@ import { assertEqual } from '../../error/assert';
 import type { QRLInternal } from '../../qrl/qrl-class';
 import type { QRL } from '../../qrl/qrl.public';
 import type { JSXOutput } from '../../render/jsx/types/jsx-node';
-import { runResource, type ResourceDescriptor } from '../../use/use-resource';
 import {
   Task,
   TaskFlags,
   cleanupTask,
   runComputed2,
-  runTask2,
+  runResource,
+  runSubscriber2,
+  type ResourceDescriptor,
   type TaskFn,
 } from '../../use/use-task';
 import { isPromise, maybeThen, maybeThenPassError, safeCall } from '../../util/promises';
@@ -301,7 +302,7 @@ export const createScheduler = (
         break;
       case ChoreType.TASK:
       case ChoreType.VISIBLE:
-        returnValue = runTask2(chore.$payload$ as Task<TaskFn, TaskFn>, container, host);
+        returnValue = runSubscriber2(chore.$payload$ as Task<TaskFn, TaskFn>, container, host);
         break;
       case ChoreType.CLEANUP_VISIBLE:
         const task = chore.$payload$ as Task<TaskFn, TaskFn>;
